`MIN_VERSION_ghc` instead of `__GLASGOW_HASKELL__`
It seems that when `ghc` tries to build the plugin with the bootstrap compiler,
`MIN_VERSION_ghc` reports that we are using the GHC API 9.10 but
`__GLASGOW_HASKELL__` reports GHC 9.8. After this change, the plugin builds
during building `ghc`.

Unfortunately, as soon as we try to build anything _with_ the plugin,
compilation fails with

```
libraries/ghc-prim/GHC/CString.hs:1:1: error: [GHC-51294]
    Failed to load interface for ‘GHC.Internal.Debug.Trace’.
    no unit id matching ‘ghc-internal’ was found
```